% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_to_unit_vec.R
\name{angle_to_unit_vec}
\alias{angle_to_unit_vec}
\title{Convert Angles to Unit Vectors}
\usage{
angle_to_unit_vec(theta)
}
\arguments{
\item{theta}{a vector of angles}
}
\value{
U a unit vector or matrix of unit vectors in each row
}
\description{
Convert Angles to Unit Vectors
}
\examples{
theta <- seq(0, 2*pi, length.out = 10)
(U <- angle_to_unit_vec(theta))
}
