% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gp_geostat_reg.R
\name{create_gp_geostat_reg_data_list}
\alias{create_gp_geostat_reg_data_list}
\title{Create Data List for Complete Data Gaussian Process Regression}
\usage{
create_gp_geostat_reg_data_list(loc1, loc2, Y, X)
}
\arguments{
\item{loc1}{Coordinates for the observed sites}

\item{loc2}{Coordinates for the unobserved sites}

\item{Y}{(vector) Observed response}

\item{X}{(matrix) Observed predictors}
}
\value{
List object to pass to cmdstan
}
\description{
Create Data List for Complete Data Gaussian Process Regression
}
\examples{
## Not run:
lst <- create_gp_geostat_reg_data_list(
 loc1 = create_grid_df(10, 10),
 loc2 = create_grid_df(15, 15),
 Y = runif(100, 0, 5),
 X = matrix(rnorm(2*100, 0, 1), nrow = 100))
## End(Not run)
}
