% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_geostat_sim_data.R
\name{create_grid_df}
\alias{create_grid_df}
\title{Create Regularly-spaced Grid of Locations}
\usage{
create_grid_df(N, M)
}
\arguments{
\item{N}{Number of columns of sites}

\item{M}{Number of rows of sites}
}
\value{
data frame with cartesian coordinates (long, lat)
}
\description{
Create Regularly-spaced Grid of Locations
}
\examples{
create_grid_df(10, 10)
}
