% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pgp_geostat_reg.R
\name{create_pgp_geostat_reg_data_list}
\alias{create_pgp_geostat_reg_data_list}
\title{Create Data List for Complete Data Projected Gaussian Process Regression}
\usage{
create_pgp_geostat_reg_data_list(loc1, loc2, theta, X)
}
\arguments{
\item{loc1}{Coordinates for the observed sites}

\item{loc2}{Coordinates for the unobserved sites}

\item{theta}{(vector) Observed angles}

\item{X}{(matrix) Observed predictors}
}
\value{
List object to pass to cmdstan
}
\description{
Create Data List for Complete Data Projected Gaussian Process Regression
}
\examples{
## Not run:
lst <- create_pgp_geostat_reg_data_list(
 loc1 = create_grid_df(10, 10),
 loc2 = create_grid_df(15, 15),
 theta = runif(100, 0, 2*pi),
 X = matrix(rnorm(2*100, 0, 1), nrow = 100))
## End(Not run)
}
