% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pn_arx_id_inc_model.R
\name{create_pnarxid_inc_data_list}
\alias{create_pnarxid_inc_data_list}
\title{Create Data List for PN ARXID Incomplete Model}
\usage{
create_pnarxid_inc_data_list(theta, X, X_ppd, sigma_0 = 100, ...)
}
\arguments{
\item{theta}{}

\item{X}{}

\item{X_ppd}{}

\item{sigma_0}{}

\item{...}{}
}
\value{
data_list a list to pass to Stan
}
\description{
Create Data List for PN ARXID Incomplete Model
}
\examples{
df <- pn_arx_sim_data(N = 100, 
  ar_X1 = c(0.24), ma_X1 = numeric(0), 
  ar_X2 = c(0.75), ma_X2 = c(-0.25, 0.25))
  s <- sample(1:90, size = 10)
  df[s, "theta"] <- NA
create_pnarxid_inc_data_list(df$theta[1:90], 
    df[1:90, c("X1", "X2")], 
    X_ppd = df[91:100, c("X1", "X2")])
}
