% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gp_geostat_reg.R
\name{fit_gp_geostat_reg_model}
\alias{fit_gp_geostat_reg_model}
\title{Fit Geostatistical Regression Model with Gaussian Process}
\usage{
fit_gp_geostat_reg_model(
  loc1,
  loc2,
  Y,
  X,
  iter_sampling = 1000,
  iter_warmup = 1000,
  refresh = 500,
  chains = 2,
  ...
)
}
\arguments{
\item{loc1}{Coordinates for the observed sites}

\item{loc2}{Coordinates for the unobserved sites}

\item{Y}{(vector) Observed response}

\item{X}{(matrix) Observed predictors}

\item{iter_sampling}{int iterations for sampling}

\item{iter_warmup}{int iterations for warmup}

\item{refresh}{int how often to print to screen}

\item{chains}{int number of MCMC chains}

\item{...}{other arguments to pass to CmdStan}
}
\value{
Cmdstan fit object
}
\description{
Fit Geostatistical Regression Model with Gaussian Process
}
\examples{
## Not run:
if (FALSE) {
  loc <- create_grid_df(10, 10)
  loc2 <- create_grid_df(11, 11)
  X <- gp_geostat_sim_data(10, 10, 1, 5, 1)$X
 
  df <- gp_geostat_sim_data(loc, X, sigma = 1, rho = 2, alpha = 0.5)
  
  df$Y <- df$X + X
  df$X <- X
         
  fit_gp_geostat_reg_model(
  loc1 = loc, loc2 = loc2,
  Y = df$Y, 
  X = df$X,
  refresh = 0,
  iter_warmup = 10,
  iter_sampling = 10,
  show_messages = FALSE,
  show_exceptions = FALSE)
  
  
}
## End(Not run)
}
