% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_jpgpmgp_reg_inc.R
\name{fit_jpgpmgp_geostat_reg_inc_model}
\alias{fit_jpgpmgp_geostat_reg_inc_model}
\title{Fit the Joint Projected Gaussian Process Multivariate Gaussian Process Geostatistical Regression Model to Incomplete Data}
\usage{
fit_jpgpmgp_geostat_reg_inc_model(
  loc1,
  loc2,
  theta,
  Y,
  X,
  iter_sampling = 1000,
  iter_warmup = 1000,
  refresh = 500,
  chains = 2,
  ...
)
}
\arguments{
\item{loc1}{Coordinates for the observed sites}

\item{loc2}{Coordinates for the unobserved sites}

\item{theta}{(vector) Incomplete observed angles}

\item{Y}{(vector) (Incomplete) observed responses}

\item{X}{(matrix) Observed predictors; should be complete}

\item{iter_sampling}{int iterations for sampling}

\item{iter_warmup}{int iterations for warmup}

\item{refresh}{int how often to print to screen}

\item{chains}{int number of MCMC chains}

\item{...}{other arguments to pass to CmdStan}
}
\value{
Cmdstan fit object
}
\description{
Fit the Joint Projected Gaussian Process Multivariate Gaussian Process Geostatistical Regression Model to Incomplete Data
}
\examples{
## Not run:
if (FALSE) {
  loc <- create_grid_df(10, 10)
  loc2 <- create_grid_df(11, 11)
  X <- gp_geostat_sim_data(10, 10, 1, 5, 1)$X
 
  df <- pgp_geostat_reg_sim_data(loc, X, sigma_w = 1, rho_w = 0, rho = 5, 
         mu0 = c(1,0), B = matrix(c(-0.5, 1), nrow = 1))
  df$Y <- cos(df$theta) - 0.2 * sin(df$theta) + rnorm(100)
  ind_miss <- 1:25 * 4
  df$theta_inc[ind_miss] <- NA
         
  fit_jpgpmgp_geostat_reg_inc_model(
  loc1 = loc, loc2 = loc2,
  theta = df$theta_inc, 
  Y = df$Y,
  X = df$X,
  refresh = 0,
  iter_warmup = 10,
  iter_sampling = 10,
  show_messages = FALSE,
  show_exceptions = FALSE)
  
  
}
## End(Not run)
}
