% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pgp_geostat.R
\name{fit_pgp_geostat_model}
\alias{fit_pgp_geostat_model}
\title{Fit the Projected Gaussian Process Geostatistical Model with Complete Data}
\usage{
fit_pgp_geostat_model(
  loc1,
  loc2,
  theta,
  iter_sampling = 1000,
  iter_warmup = 1000,
  refresh = 500,
  chains = 2,
  ...
)
}
\arguments{
\item{loc1}{Coordinates for the observed sites}

\item{loc2}{Coordinates for the unobserved sites}

\item{theta}{vector of angles}

\item{iter_sampling}{int iterations for sampling}

\item{iter_warmup}{int iterations for warmup}

\item{refresh}{int how often to print to screen}

\item{chains}{int number of MCMC chains}

\item{...}{other arguments to pass to CmdStan}
}
\value{
Cmdstan fit object
}
\description{
Fit the Projected Gaussian Process Geostatistical Model with Complete Data
}
\examples{
## Not run:
if (FALSE) {
  loc <- create_grid_df(10, 10)
  loc2 <- create_grid_df(11, 11)
 
  df <- pgp_geostat_sim_data(loc, sigma_w = 1, rho_w = 0, rho = 5, 
         mu = c(1,0))
         
  fit_pgp_geostat_model(
  loc1 = loc, loc2 = loc2,
  theta = df$theta,
  refresh = 0,
  iter_warmup = 10,
  iter_sampling = 10,
  show_messages = FALSE,
  show_exceptions = FALSE)
  
  
}
## End(Not run)
}
