% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pn_arx_id_inc_model.R
\name{fit_pn_arx_id_inc_model}
\alias{fit_pn_arx_id_inc_model}
\title{Fit the Projected Normal ARX Process with Incomplete Angular Data}
\usage{
fit_pn_arx_id_inc_model(
  theta,
  X,
  X_ppd,
  iter_sampling = 1000,
  iter_warmup = 1000,
  refresh = 500,
  chains = 2,
  ...
)
}
\arguments{
\item{theta}{}

\item{X}{}

\item{X_ppd}{}

\item{iter_sampling}{}

\item{iter_warmup}{}

\item{refresh}{}

\item{chains}{}

\item{...}{}
}
\value{
CmdStan fit
}
\description{
Fit the Projected Normal ARX Process with Incomplete Angular Data
}
\examples{
## Not run:
if (instantiate::stan_cmdstan_exists()) {
  df <- pn_arx_sim_data(N = 100, 
  ar_X1 = c(0.24), ma_X1 = numeric(0), 
  ar_X2 = c(0.75), ma_X2 = c(-0.25, 0.25))
  s <- sample(1:90, size = 10)
  df[s, "theta"] <- NA
  
  df_ppd <- df[91:100,]
  fit_pn_arx_id_inc_model(theta = df$theta[1:90], 
  X = df[1:90,c("X1", "X2")], 
  X_ppd = df_ppd[,c("X1", "X2")])
  
  
}
## End(Not run)
}
