% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pn_arx_id_model.R
\name{fit_pn_arx_id_model}
\alias{fit_pn_arx_id_model}
\title{Projected Normal AR Process with Exogenous Predictors (Identity Covariance)}
\usage{
fit_pn_arx_id_model(
  theta,
  X,
  X_ppd,
  iter_sampling = 1000,
  iter_warmup = 1000,
  refresh = 500,
  chains = 2,
  ...
)
}
\arguments{
\item{theta}{vector angles}

\item{X}{matrix of exogenous predictors}

\item{X_ppd}{matrix of exogenous predictors for PPD draws}

\item{iter_sampling}{int iterations for sampling}

\item{iter_warmup}{int iterations for warmup}

\item{refresh}{int how often to print to screen}

\item{chains}{int number of MCMC chains}

\item{...}{other arguments to pass to CmdStan}
}
\value{
vector of ppd drawn angles
}
\description{
The model is fitted using the projected normal regression with identity covariance.
}
\examples{
## Not run:
if (instantiate::stan_cmdstan_exists()) {
  df <- pn_arx_sim_data(N = 100, 
  ar_X1 = c(0.24), ma_X1 = numeric(0), 
  ar_X2 = c(0.75), ma_X2 = c(-0.25, 0.25))
  
  df_ppd <- df[91:100,]
  fit_pn_arx_model(theta = df$theta[1:90], 
  X = df[1:90,c("X1", "X2")], 
  X_ppd = df_ppd[,c("X1", "X2")])
}
## End(Not run)
}
