% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pn_arx_model.R
\name{fit_pn_arx_model}
\alias{fit_pn_arx_model}
\title{Title}
\usage{
fit_pn_arx_model(
  theta,
  X,
  X_ppd,
  iter_sampling = 1000,
  iter_warmup = 1000,
  refresh = 500,
  chains = 2,
  ...
)
}
\arguments{
\item{theta}{vector angular time series; values should be between 0 and 2*pi}

\item{X}{vector or matrix of predictor time series for training data}

\item{X_ppd}{vector or matrix of predictor time series for testing data}

\item{iter_sampling}{int MCMC sampling iterations to send to CmdStan}

\item{iter_warmup}{int MCMC warmup iterations to send to CmdStan}

\item{refresh}{int how often should CmdStan print an update during sampling}

\item{chains}{int number of MCMC chains}

\item{...}{additional arguments to pass to CmdStan}
}
\value{
fit a CmdStan fit object
}
\description{
Title
}
\examples{
## Not run:
if (instantiate::stan_cmdstan_exists()) {
  df <- pn_arx_sim_data(N = 100, 
  ar_X1 = c(0.24), ma_X1 = numeric(0), 
  ar_X2 = c(0.75), ma_X2 = c(-0.25, 0.25))
  
  df_ppd <- df[91:100,]
  fit_pn_arx_model(theta = df$theta[1:90], 
  X = df[1:90,c("X1", "X2")], 
  X_ppd = df_ppd[,c("X1", "X2")])
}
## End(Not run)
}
