% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit_pnreg_identity_model}
\alias{fit_pnreg_identity_model}
\title{Fit the Projected Normal Regression model with Identity Covariance.}
\usage{
fit_pnreg_identity_model(theta, X, X_ppd, ...)
}
\arguments{
\item{theta}{a vector length N of angles}

\item{X}{a matrix of size N x P of predictors}

\item{X_ppd}{a matrix of size N_tilde x P of predictors for the posterior predictive draws}
}
\value{
A data frame of posterior summaries.
}
\description{
Fit the PN Regression Stan model and return posterior summaries.
}
\examples{
if (instantiate::stan_cmdstan_exists()) {
  fit_pnreg_identity_model(theta, X, X_ppd)
}
}
\concept{models}
