% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_geostat_sim_data.R
\name{gp_geostat_sim_data}
\alias{gp_geostat_sim_data}
\title{Simulate from a GP Geostatistical Model}
\usage{
gp_geostat_sim_data(
  N,
  M,
  sigma,
  rho,
  alpha,
  iter_sampling = 1000,
  iter_warmup = 1000,
  refresh = 0,
  chains = 2,
  ...
)
}
\arguments{
\item{N}{Number of columns of sites}

\item{M}{Number of rows of sites}

\item{sigma}{Standard deviation of the process}

\item{rho}{Length scale parameter for the exponential covariance function (inverse of spatial decay)}

\item{alpha}{Magnitude parameter for the exponential covariance function}

\item{iter_sampling}{int iterations for sampling}

\item{iter_warmup}{int iterations for warmup}

\item{refresh}{int how often to print to screen}

\item{chains}{int number of MCMC chains}

\item{...}{other arguments to pass to CmdStan}
}
\value{
Data frame with simulated GP and location coordinates
}
\description{
Simulate from a GP Geostatistical Model
}
\examples{
## Not run:
if (instantiate::stan_cmdstan_exists()) {
  loc <- create_grid_df(10, 10)
  loc2 <- create_grid_df(11, 11)
 
  df <- gp_geostat_sim_data(N = 10, M = 10, sigma = 1, rho = 10, alpha = 5)   
  
}
## End(Not run)
}
