% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgp_geostat_reg_sim_data.R
\name{pgp_geostat_reg_sim_data}
\alias{pgp_geostat_reg_sim_data}
\title{Simulate from Projected Gaussian Process Geostatistical Regression Model}
\usage{
pgp_geostat_reg_sim_data(
  loc,
  X,
  sigma_w,
  rho_w,
  rho,
  mu0,
  B,
  iter_sampling = 1000,
  iter_warmup = 1000,
  refresh = 0,
  chains = 2,
  ...
)
}
\arguments{
\item{loc}{Matrix of coordinates of N sites.}

\item{X}{Predictor matrix (N x P)}

\item{sigma_w}{Standard deviation of the first latent angular coordinate}

\item{rho_w}{Correlation between latent angular coordinates}

\item{rho}{Length scale parameter of the exponential covariance function}

\item{mu0}{Latent bivariate intercept vector}

\item{B}{Regression coefficient matrix (P x 2)}

\item{iter_sampling}{int iterations for sampling}

\item{iter_warmup}{int iterations for warmup}

\item{refresh}{int how often to print to screen}

\item{chains}{int number of MCMC chains}

\item{...}{other arguments to pass to CmdStan}
}
\value{
Data frame with simulated PGP, predictors, and location coordinates
}
\description{
Simulate from Projected Gaussian Process Geostatistical Regression Model
}
\examples{
## Not run:
if (FALSE) {
  loc <- create_grid_df(10, 10)
  loc2 <- create_grid_df(11, 11)
  X <- rnorm(100)
  B <- matrix(c(1, 2), nrow = 1)
 
  df <- pgp_geostat_reg_sim_data(loc, X = X, sigma_w = 1, rho_w = 0, rho = 5,
         mu0 = c(1,0), B = B)   
  
}
## End(Not run)
}
