% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgp_geostat_sim_data.R
\name{pgp_geostat_sim_data}
\alias{pgp_geostat_sim_data}
\title{Simulate from Projected Gaussian Process Geostatistical Model}
\usage{
pgp_geostat_sim_data(
  N,
  M,
  sigma_w,
  rho_w,
  rho,
  mu,
  iter_sampling = 1000,
  iter_warmup = 1000,
  refresh = 0,
  chains = 2,
  ...
)
}
\arguments{
\item{N}{Number of columns of sites}

\item{M}{Number of rows of sites}

\item{sigma_w}{Standard deviation of the first latent angular coordinate}

\item{rho_w}{Correlation between latent angular coordinates}

\item{rho}{Length scale parameter of the exponential covariance function}

\item{mu}{Latent bivariate normal mean vector}

\item{iter_sampling}{int iterations for sampling}

\item{iter_warmup}{int iterations for warmup}

\item{refresh}{int how often to print to screen}

\item{chains}{int number of MCMC chains}

\item{...}{other arguments to pass to CmdStan}
}
\value{
Data frame with simulated PGP and location coordinates
}
\description{
Simulate from Projected Gaussian Process Geostatistical Model
}
\examples{
## Not run:
if (FALSE) {
  loc <- create_grid_df(10, 10)
  loc2 <- create_grid_df(11, 11)
 
  df <- pgp_geostat_sim_data(N, M, sigma_w = 1, rho_w = 0, rho = 5, 
         mu = c(1,0))   
  
}
## End(Not run)
}
