% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pn_arx_sim_data.R
\name{pn_arx_sim_data}
\alias{pn_arx_sim_data}
\title{Simulate Projected Normal Data from a latent ARX process}
\usage{
pn_arx_sim_data(
  N = 100,
  mu_0 = c(0, 0),
  B_vec = c(1, 0.3, 0, -2.5),
  Psi_vec = c(0.75, -0.2, -0.2, 0.5),
  Sigma_vec = c(1, 0, 0, 1),
  X = NULL,
  ar_X1 = NULL,
  ma_X1 = NULL,
  ar_X2 = NULL,
  ma_X2 = NULL
)
}
\arguments{
\item{N}{int time series length}

\item{mu_0}{initial latent mean vector}

\item{B_vec}{vector of regression coefficients}

\item{Psi_vec}{vector of autoregressive coefficients}

\item{Sigma_vec}{vector of stationary covariance parameters}

\item{X}{NULL or matrix of predictor time series}

\item{ar_X1}{NULL or vector of autoregressive coefficients for simulated X1}

\item{ma_X1}{NULL or vector of moving average coefficients for simulated X1}

\item{ar_X2}{NULL or vector of autoregressive coefficients for simulated X2}

\item{ma_X2}{NULL or vector of moving average coefficients for simulated X2}
}
\value{
df a data frame with theta, U1, U2, and either simulated X time series or provided X_mat time series
}
\description{
Simulate Projected Normal Data from a latent ARX process
}
\examples{
df <- pn_arx_sim_data(N = 100, ar_X1 = c(0.24), ma_X1 = numeric(0),
                      ar_X2 = c(0.75), ma_X2 = c(-0.25, 0.25))
}
