% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnreg_sim_data.R
\name{pnreg_sim_data}
\alias{pnreg_sim_data}
\title{Simulate Projected Normal Regression Data}
\usage{
pnreg_sim_data(
  N = 100,
  B = c(5, 0.5, 0, 0.25),
  Sigma_theta = c(1, 0, 0, 1),
  X = NULL,
  mu_X = NULL,
  sigma_X = NULL
)
}
\arguments{
\item{N}{Sample size}

\item{B}{Regression coefficients passed as a vector turned into a matrix (Px2) by row}

\item{Sigma_theta}{Projected normal covariance values passed as a vector turned into a matrix (2x2) by row, should be symmetric}

\item{X}{a matrix (or data frame) of (N x P) predictors or NULL}

\item{mu_X}{numeric for mean of the distribution of a single simulated predictor X or NULL}

\item{sigma_X}{non-negative numeric for the standard deviation of the single simulated predictor X or NULL}
}
\value{
df A data frame with three columns for theta, U1, U2, and columns for X (1 col simulated or the passed X appended)
}
\description{
Simulate Projected Normal Regression Data
}
\examples{
 df <- pnreg_sim_data(N = 100, mu_X = 0, sigma_X = 1)
}
