% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pnreg_gen_model.R
\name{fit_pnreg_gen_model}
\alias{fit_pnreg_gen_model}
\title{Fit the Projected Normal Regression model with General Covariance.}
\usage{
fit_pnreg_gen_model(
  theta,
  X,
  X_ppd,
  iter_sampling = 1000,
  iter_warmup = 3000,
  refresh = 500,
  ...
)
}
\arguments{
\item{theta}{a vector length N of angles}

\item{X}{a matrix of size N x P of predictors}

\item{X_ppd}{a matrix of size N_tilde x P of predictors for the posterior predictive draws}

\item{iter_sampling}{int number of iterations to sample per chain}

\item{iter_warmup}{int number of warmup iterations to sample per chain}

\item{refresh}{int how often CmdStan should print an update. Set to 0 to suppress updates.}

\item{...}{Other arguments to pass to CmdStan}
}
\value{
A data frame of posterior summaries.
}
\description{
Fit the PN Regression Stan model and return posterior summaries. Covariance follows the constrained parameterization from
}
\examples{
## Not run:
if (instantiate::stan_cmdstan_exists()) {
  df <- pnreg_sim_data(N = 100, mu_X = 0, sigma_X = 1)
  fit_pnreg_identity_model(theta = df$theta, X = df$X, X_ppd = df$X,
  refresh = 0,
  iter_warmup = 10,
  iter_sampling = 10,
  show_messages = FALSE,
  show_exceptions = FALSE)
}
## End(Not run)
}
\seealso{
Other models: 
\code{\link{fit_pnreg_identity_model}()},
\code{\link{fit_vm_brms_model}()}
}
\concept{models}
